pyproject.toml
[project]
name = "patientadvocacy"
version = "0.1.0"
description = "Patient advocacy assistant built with Google ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.16.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=0.3.0",
    "openai-whisper>=20231117",
    "python-multipart>=0.0.9",
    "litellm>=1.79.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/patientadvocacy"]

